+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/model/Path.kt
@@ -64,6 +64,13 @@ data class Path(val edgesList: List<String>) {
             if (isSingle) Path.trivialPath()
             else Path(edgesList.drop(1))
 
+    /**
+     * @return parent, i.e. the path without the last edge, if not trivial, trivial element if trivial
+     */
+    val parent: Path
+        get() =
+            if (isSingle) Path.trivialPath()
+            else Path(edgesList.dropLast(1))
 
     /**
      * @param path that will be added after the present path